api/requirements.txt
azure-functions
azure-cosmos==4.7.0